Move unflag-billed services
