+++ app/services/bill-runs/send/update-invoice-numbers.service.js
@@ -11,7 +11,7 @@ const ExpandedError = require('../../../errors/expanded.error.js')
 const { calculateAndLogTimeTaken, timestampForPostgres } = require('../../../lib/general.lib.js')
 const ChargingModuleSendBillRunRequest = require('../../../requests/charging-module/send-bill-run.request.js')
 const ChargingModuleViewBillRunRequest = require('../../../requests/charging-module/view-bill-run.request.js')
-const UnflagBilledLicencesService = require('../supplementary/unflag-billed-licences.service.js')
+const UnflagBilledSupplementaryLicencesService = require('../unflag-billed-supplementary-licences.service.js')
 
 /**
  * Updates a bill run and its bill records with the invoice numbers generated by the Charging Module API
@@ -41,7 +41,7 @@ async function go(billRun) {
     await _updateBillRun(billRunId, externalBillRun)
 
     if (billRun.batchType === 'supplementary') {
-      await UnflagBilledLicencesService.go(billRun)
+      await UnflagBilledSupplementaryLicencesService.go(billRun)
     }
 
     calculateAndLogTimeTaken(startTime, 'Send bill run complete', { billRun })
